requirements.txt
fastapi==0.52.0
line-bot-sdk==1.15.0
uvicorn==0.10.9
